Moved sse stuff under sse api
